+++ internal/backends/winit/event_loop.rs
@@ -378,11 +378,19 @@ fn process_window_event(
             let ev = match touch.phase {
                 winit::event::TouchPhase::Started => {
                     *pressed = true;
-                    MouseEvent::Pressed { position, button: PointerEventButton::Left, click_count: 0 }
+                    MouseEvent::Pressed {
+                        position,
+                        button: PointerEventButton::Left,
+                        click_count: 0,
+                    }
                 }
                 winit::event::TouchPhase::Ended | winit::event::TouchPhase::Cancelled => {
                     *pressed = false;
-                    MouseEvent::Released { position, button: PointerEventButton::Left, click_count: 0 }
+                    MouseEvent::Released {
+                        position,
+                        button: PointerEventButton::Left,
+                        click_count: 0,
+                    }
                 }
                 winit::event::TouchPhase::Moved => MouseEvent::Moved { position },
             };
