+++ internal/compiler/builtins.slint
@@ -90,7 +90,7 @@ export struct PointerEvent := {
     //-name:slint::private_api::PointerEvent
     button: PointerEventButton,
     kind: PointerEventKind,
-    click_count: int,
+    click-count: int,
 }
 
 export TouchArea := _ {
